Refactor navbar rendering and show errors in separate page
Put navbar rendering in a separate function called -navbar that shows all
nav links but automatically makes the current page a non-link.

Show full error information in a separate page, and make the error indicator
on the sources page show right next to the source directory, rather than off
to the side in a separate div.