Remove *letmap-omitted-key* feature from letmap
This is no longer in use - it was being used in order to look at private data
in objects in tests. We are no longer enforcing encapsulation this way, so
a hook to re-expose this data is no longer necessary. We are still using :omit
in letmap, but only for storing trivial intermediate values.