Change hesokuri.source to be amenable to multiple live-edit branches
Before, hesokuri.source assumed the live-edit branch was a single constant
string. Now it uses a function to determine if a given branch name is live-edit.
The advance-a function changed the most. It become slightly clearer because the
optimizations that were in there before had to be removed (and probably should
never have been there in the first place).

Also, when pushing, we no longer try to push local versions of live-edit
branches "normally" before forcing; we always force, pushing to an underscored
name like foo_hesokr_bar. This is because once the live edit branch name is
configurable, there is no guarantee that the other peers agree what branches
qualify as live-edit.