Move validation logic of source-def/kind into a separate function
New function is source-def/validation-error, which returns a description of the
problem if there is one, or nil if there is none. This allows the heso object to
explicitly check if the configuration is valid, rather than allowing an
exception to occur if it is not.

Checks both the :live-edit-branches (if present) and :host-to-path entries.