Add an unwanted-branch? function to source-def
To help clean-up obsolete branches, or to delete branches without worrying
about another peer pushing the branch back to you, we'll add an
:unwanted-branches entry to source-def, which is an optional entry that contains
a set of branches to delete automatically. Details are in the
hesokuri.source-def namespace docstring.

This change does not cause the setting to be honored in the source object. It
merely allows it to be read.