Make the watcher object white-box
Stop using anonymous functions and stop hiding arguments to callbacks inside of
closures. This should make the dump more informative and help with debugging. It
has made the code a little more readable and linear.

Additionally, merge the types of watchers so that watchers for dirs and files
look the same. The one for files simply delegates to dirs.