Implement "see" module which abbreviates objects in a data structure
This is used now in the web module for dumping the state of the app. This allows
us to include more information in object state without worrying about the time
needed to generate the pretty-printed state. With this module, showing a dump is
much faster and it can include self-nested objects.