Add dependencies on barbarywatchservice and jna libraries
barbarywatchservice supplies a file system watcher similar to that in
java.nio.file. However, java.nio.file uses polling to watch for changes on Mac
OS X. barbarywatchservice uses native OS X APIs to listen to filesystem changes.

barbarywatchservice does not seem to be available in Maven or Clojars, so we
add the source to this repository. It is compiled by Leiningen automatically.

We also must pull in the jna library from Maven, since it is required to compile
barbarywatchservice.

This commit does not actually cause watcher.clj to use barbarywatchservice. That
will come in a follow-up CL.