Fix hesokuri.repo/init behavior when initializing a bare repo
There was a bug where it would initialize a non-bare repo with a working
directory equal to the git directory of the bare repo. This was a problem both
when the bare repo was already in that location, and when there was not repo
there but the :bare property was set to true.

Also add a test case for proper creation of non-bare repos in init.