Make value functions keep their name and fields
Fields are the values of the pre-supplied arguments. We are now using proxy to
implement the AFunction and VFN interfaces. We need to use proxy for more
reasons eventually, including to override equals and hashCode, so we do this
rather than just using (fn ...) and extendsing its class to a protocol.