Change the way we confirm success for creating temp repos
This change allows Windows machines to pass hesokuri.git-test. Before this
change, we looked for the .git directory in the output of git init to confirm
everything worked. However, msysgit outputs the path in a different form than
what is normal (/ instead of \, e.g. "C:/Users/JDoe/foo"). So now, just use
throw-if-error to confirm success.

Note that searching for "Initialized empty Git repository", though tempting,
does not work for localized environments.