Allow hesokuri.ssh connections to use multiple sessions
The plan is for each session to host a single hesoprot command, so that multiple
commands do not require re-authenticating. By using the SSH session in this
manner, we don't have to manually escape or delimit data generated by something
like git-receive-pack and git-send-pack. We can just attach the session streams
to the STDIO of these commands, and close the session when the commands are
finished executing.