Add work-tree function to hesokuri.git/Context
This lets the context tell you what the work tree is. The default behavior is
to remove the ".git" path segment from the git directory, if the directory path
ends with "/.git"

Also, extend the Context protocol to Map. This lets us tack on extra data to
almost any data structure for allowing it to be used as a Git context.