Add normalization functionality to source-def module
Normalization removes legacy constructs and replaces them with the most modern
construct. This will be necessary to support automatic editing of
configurations.

There is now a way to specify "all/any commits" for an unwanted branch when
using a map for :unwanted-branches, since this is possible with a set, and the
"normlized" form is a map.